requirements.txt
streamlit
db-dtypes
google-cloud-bigquery
openpyxl
xlsxwriter
imagekitio
openai
tiktoken
python-barcode
fpdf
matplotlib
urllib3<2
db_dtypes
gspread
oauth2client
google-cloud-storage
google-cloud-bigquery
google-api-python-client
streamlit-oauth
keepa
pandas_gbq
pinecone
gdown
pyotp